fix: throttle lazy.nvim git ops
I've recently experienced issues with lazy fetching/checking out git
repos where my ssh-agent was refused to accept requests. It appears to
be an issue with bandwidth as reducing the amount of requests at a time
allows for successful lazy operations. This will require further
investigation as I never had an issue with this behavior before.

Signed-off-by: wrightbradley <bradley.wright.tech@gmail.com>